% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxer.R
\name{boxer}
\alias{boxer}
\title{Create a bounding box in an image based on selected points}
\usage{
boxer(names, file_path_input, file_path_output, color = "red",
  size_x = NA, size_y = NA, classifier = NA, show_classifier = F)
}
\arguments{
\item{names}{A list that contains the names of the image files to cycle
through.}

\item{file_path_input}{The directory that contains the image files in JPEG
format.}

\item{file_path_output}{The directory to which the annotated images should be
saved.}

\item{color}{The color of the bounding box. Default is "red".}

\item{size_x}{Number of columns to resize the x-axis. Default is NA and will
return size of original image.}

\item{size_y}{Number of columns to resize the y-axis. Default is NA and will
return size of original image.}

\item{classifier}{Character string to add a classifier along with the output
coordinates. Default is NA.}

\item{show_classifier}{Logical to indicate to print the classifier along with
the bounding box. Default is F.}
}
\value{
A dataframe with the bounding box coordinates for each image, and a
  new image file with the bounding box annotation for each of the images
  processed.
}
\description{
\code{boxer} allows the user to identify an area in an image around which to
generate a bounding box.
}
\details{
Stream a series of JPEG or PNG images from a directory that each appears on
the screen, and indicate between 2 and 26 points that define an area around
which a bounding box is generated based on the max/min x and y values of the
selected coordinates. There are two outputs. One is a dataframe with the name
of the original image, the coordinates of the bounding box ("x_left",
"y_top", "x_right", "y_bottom") and the classifier. Another is, for each
image, an image file "out_<original_file_name>" annotated with the bounding
box, the classifier printed above the bounding box if show_classifer is set
to TRUE, and resized if either the size_x or size_y parameter is given.
}
\examples{
\dontrun{
boxer(names = c("dog.275.jpg", "dog.544.jpg", "dog.119.jpg"),
file_path_input = "/dog_pics/input/", file_path_output = "/dog_pics/output/",
color = "blue", size_x = 224, size_y = 224, classifier = "dog",
show_classifier = T)
}

}
